requirement.txt
mysql-connector-python==8.0.33
bcrypt==4.0.1
reportlab==4.0.4
openpyxl==3.1.2
pillow==12.1.1
python-dateutil==2.8.2
pyinstaller==6.19.0